pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qernel"
version = "0.0.1.dev"
description = "An agentic, virtual kernel for efficient programming of quantum systems"
readme = "README.md"
authors = [
    {name = "Seun Omonije"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "cirq-core>=1.0",
    "cirq-web>=1.0",
    "stim>=1.12",
    "pyyaml>=6.0",
    "requests>=2.25.0",
    "typing-extensions>=4.0.0",
    "pywebview>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["qernel*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]
